\name{RcppPrjTmpl-package}
\alias{RcppPrjTmpl-package}
\alias{RcppPrjTmpl}
\docType{package}
\title{
    \packageTitle{RcppPrjTmpl}
}
\description{
    \packageDescription{RcppPrjTmpl}
}
\details{
    \packageDESCRIPTION{RcppPrjTmpl}
    \packageIndices{RcppPrjTmpl}
    \preformatted{}
}
\author{
    \packageAuthor{RcppPrjTmpl}
    Maintainer: \packageMaintainer{RcppPrjTmpl}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
